soy-api2/requirements.txt
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.6
onnxruntime==1.16.0
numpy==1.26.0
pillow==10.0.1
pandas==2.1.1
scikit-image==0.22.0
python-dotenv==1.0.0
pydantic==2.4.2
scipy==1.11.3
matplotlib==3.8.0
opencv-python==4.11.0.86
ultralytics==8.3.158
# backports.tarfile==1.2.0
# pyinstaller==6.14.1